# openapi/marketflow-orders-api.yaml
openapi: 3.0.3

info:
  title: MarketFlow Orders API
  version: 0.4.0
  description: >
    Multi-seller MVP. Prices are snapshotted at creation and stock is debited at
    payment. Each seller ships their own part. The overall order completes only
    after the buyer confirms receipt of every part. Paid orders can be refunded
    before a seller starts fulfillment; pending seller and platform accruals are reversed.

servers:
  - url: http://localhost:8080

tags:
  - name: Orders
  - name: Seller fulfillment

security:
  - sessionCookie: []

paths:
  /api/v1/orders:
    get:
      tags: [Orders]
      summary: List the authenticated buyer's orders
      operationId: listBuyerOrders
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Orders returned newest first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPage'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Orders]
      summary: Create an order from selected cart items
      description: >
        Saves product names and prices as snapshots, creates one seller part for
        every seller, and removes the selected cart items. Stock is not debited
        until payment.
      operationId: createOrder
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '201':
          description: Order created
          headers:
            Location:
              schema: { type: string, format: uri-reference }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedOrder'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/v1/orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get the buyer's own order and item snapshots
      operationId: getOrder
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetails'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'

  /api/v1/orders/{orderId}/history:
    get:
      tags: [Orders]
      summary: Get the event history of the buyer's own order
      description: Returns asynchronously recorded RabbitMQ events in chronological order.
      operationId: getOrderHistory
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order event history, oldest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderHistoryView'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'

  /api/v1/orders/{orderId}/cancel:
    post:
      tags: [Orders]
      summary: Cancel an unpaid order or refund a paid order before fulfillment
      operationId: cancelOrder
      parameters:
        - $ref: '#/components/parameters/OrderId'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: Order cancelled or refunded; repeating cancellation is a no-op
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/v1/orders/{orderId}/fulfillments:
    get:
      tags: [Orders]
      summary: Get the seller parts of the buyer's own order
      operationId: getOrderFulfillments
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Seller parts returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SellerOrder'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'

  /api/v1/orders/{orderId}/fulfillments/{partId}/receive:
    post:
      tags: [Orders]
      summary: Confirm receipt of one shipped seller part
      description: >
        Marks the part as USERGETPRODUCT and transactionally stores a RELEASE_SELLER_FUNDS
        outbox command. Funds remain in pending balances until the RabbitMQ settlement consumer
        moves seller funds and platform commission to available balances.
      operationId: receiveSellerOrder
      parameters:
        - $ref: '#/components/parameters/OrderId'
        - $ref: '#/components/parameters/PartId'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: Receipt confirmed and release command queued; repeating it is a no-op
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/v1/seller/orders:
    get:
      tags: [Seller fulfillment]
      summary: List paid order parts belonging to the authenticated seller
      operationId: listSellerOrders
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Seller parts returned newest first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellerOrderPage'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/v1/seller/orders/{partId}:
    get:
      tags: [Seller fulfillment]
      summary: Get the seller's own part and its item snapshots
      operationId: getSellerOrder
      parameters:
        - $ref: '#/components/parameters/PartId'
      responses:
        '200':
          description: Seller part returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellerOrderDetails'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'

  /api/v1/seller/orders/{partId}/process:
    post:
      tags: [Seller fulfillment]
      summary: Move the seller's own part from NEW to PROCESSING
      operationId: processSellerOrder
      parameters:
        - $ref: '#/components/parameters/PartId'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: Processing started; repeating the current state is a no-op
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/v1/seller/orders/{partId}/ship:
    post:
      tags: [Seller fulfillment]
      summary: Move the seller's own part from PROCESSING to SELLERSENDPRODUCT
      operationId: shipSellerOrder
      parameters:
        - $ref: '#/components/parameters/PartId'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: Part shipped; repeating the current state is a no-op
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'

components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID

  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    OrderId:
      name: orderId
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
    PartId:
      name: partId
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
    CsrfToken:
      name: X-CSRF-TOKEN
      in: header
      required: true
      description: Token returned by GET /api/v1/csrf for the same HTTP session
      schema: { type: string }

  schemas:
    CreatedOrder:
      type: object
      required: [orderId]
      properties:
        orderId: { type: integer, format: int64, minimum: 1 }

    OrderSummary:
      type: object
      required: [id, status, paymentStatus, totalPrice, createdAt]
      properties:
        id: { type: integer, format: int64, minimum: 1 }
        status:
          $ref: '#/components/schemas/OrderStatus'
        paymentStatus:
          $ref: '#/components/schemas/PaymentStatus'
        totalPrice: { type: number, multipleOf: 0.01 }
        createdAt: { type: string, format: date-time }

    OrderPage:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/OrderSummary'
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 100 }
        totalElements: { type: integer, format: int64, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }

    OrderDetails:
      type: object
      required: [id, status, paymentStatus, totalPrice, createdAt, items]
      properties:
        id: { type: integer, format: int64, minimum: 1 }
        status:
          $ref: '#/components/schemas/OrderStatus'
        paymentStatus:
          $ref: '#/components/schemas/PaymentStatus'
        totalPrice: { type: number, multipleOf: 0.01 }
        createdAt: { type: string, format: date-time }
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'

    OrderHistoryView:
      type: object
      required: [eventId, eventType, description, occurredAt]
      properties:
        eventId: { type: string, format: uuid }
        eventType:
          $ref: '#/components/schemas/MarketFlowEventType'
        description: { type: string, minLength: 1 }
        sellerOrderId: { type: integer, format: int64, minimum: 1, nullable: true }
        previousStatus: { type: string, nullable: true }
        currentStatus: { type: string, nullable: true }
        occurredAt: { type: string, format: date-time }

    OrderItem:
      type: object
      required: [productId, sellerId, productName, unitPrice, quantity, totalPrice]
      properties:
        productId: { type: integer, format: int64, minimum: 1 }
        sellerId: { type: integer, format: int64, minimum: 1 }
        productName: { type: string, minLength: 1, maxLength: 200 }
        unitPrice: { type: number, multipleOf: 0.01 }
        quantity: { type: integer, format: int32, minimum: 1 }
        totalPrice: { type: number, multipleOf: 0.01 }
        imageUrl: { type: string, nullable: true, maxLength: 1000 }

    SellerOrder:
      type: object
      required: [id, orderId, sellerId, status, totalAmount, commissionRate, commissionAmount, sellerAmount, settlementStatus, createdAt]
      properties:
        id: { type: integer, format: int64, minimum: 1 }
        orderId: { type: integer, format: int64, minimum: 1 }
        sellerId: { type: integer, format: int64, minimum: 1 }
        status:
          $ref: '#/components/schemas/OBSERFFORSENDBYSELLERPRODUCTSTATUS'
        totalAmount: { type: number, multipleOf: 0.01 }
        commissionRate: { type: number, minimum: 0, maximum: 0.9999 }
        commissionAmount: { type: number, minimum: 0, multipleOf: 0.01 }
        sellerAmount: { type: number, minimum: 0, multipleOf: 0.01 }
        settlementStatus:
          $ref: '#/components/schemas/OBSERFFORSENDFROMUSERMONEYINSELLERSTATUS'
        createdAt: { type: string, format: date-time }
        shippedAt: { type: string, format: date-time, nullable: true }
        deliveredAt: { type: string, format: date-time, nullable: true }

    SellerOrderPage:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/SellerOrder'
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 100 }
        totalElements: { type: integer, format: int64, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }

    SellerOrderDetails:
      type: object
      required: [fulfillment, paymentStatus, items]
      properties:
        fulfillment:
          $ref: '#/components/schemas/SellerOrder'
        paymentStatus:
          $ref: '#/components/schemas/PaymentStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'

    OrderStatus:
      type: string
      enum: [CREATED, CONFIRMED, SELLERSSTARTWORK, SELLERSENDWORKANDSEND, COMPLETED, CANCELLED]

    PaymentStatus:
      type: string
      enum: [NOT_PAID, PROCESSING, PAID, FAILED, REFUNDED]

    MarketFlowEventType:
      type: string
      enum:
        - ORDER_CREATED
        - ORDER_PAID
        - ORDER_CANCELLED
        - ORDER_COMPLETED
        - ORDER_REFUNDED
        - SELLER_STARTED_WORK
        - SELLER_SENT_PRODUCT
        - USER_RECEIVED_PRODUCT
        - SELLER_MONEY_PENDING
        - PLATFORM_COMMISSION_ADDED
        - SELLER_MONEY_AVAILABLE
        - SELLER_WITHDRAWAL_COMPLETED
        - SELLER_MONEY_RETURNED
        - PLATFORM_COMMISSION_RETURNED

    OBSERFFORSENDFROMUSERMONEYINSELLERSTATUS:
      type: string
      enum: [NOT_DISTRIBUTE, PENDINGWALLET, MAINWALLET, RETURNMONEY]

    OBSERFFORSENDBYSELLERPRODUCTSTATUS:
      type: string
      description: RETURNED is retained only for historical rows
      enum: [NEW, PROCESSING, SELLERSENDPRODUCT, USERGETPRODUCT, CANCELLED, RETURNED]

    ApiError:
      type: object
      required: [timestamp, status, code, message, path, validationErrors]
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer, format: int32 }
        code: { type: string }
        message: { type: string }
        path: { type: string }
        validationErrors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        traceId: { type: string, nullable: true }

    ValidationError:
      type: object
      required: [field, message]
      properties:
        field: { type: string }
        message: { type: string }

  responses:
    InvalidRequest:
      description: An ID, page parameter, quantity, or stored cart value is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    AuthenticationRequired:
      description: Authentication is required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Role, ownership, or CSRF validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ResourceNotFound:
      description: Resource does not exist or does not belong to this user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Product availability or the current status prevents the operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'